split type checking from error annotator
