+++ src/main/kotlin/org/move/lang/core/types/ty/Ty.kt
@@ -50,3 +50,5 @@ interface Ty : TypeFoldable<Ty> {
 
     fun abilities(): Set<Ability>
 }
+
+val Ty.isAbilitiesType: Boolean get() = this is TyInfer || this is TyTypeParameter || this is TyStruct
