+++ src/main/resources/inspectionDescriptions/MvMissingAcquires.html
@@ -1,0 +1,5 @@
+<html>
+<body>
+<p>Missing acquires</p>
+</body>
+</html>
