+++ src/main/resources/inspectionDescriptions/MvTypeCheck.html
@@ -1,0 +1,5 @@
+<html>
+<body>
+<p>Type checking</p>
+</body>
+</html>
